pyproject.toml
[project]
name = "carcare"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "django>=5.2.7",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-stubs>=3.16.4",
    "google-genai>=1.46.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "supabase>=2.27.0",
    "types-colorama>=0.4.15.20250801",
    "types-psycopg2>=2.9.21.20251012",
]
